feat: wrapup component family tool
