+++ src/deepset_mcp/tools/haystack_service.py
@@ -1,0 +1,31 @@
+from deepset_mcp.api.exceptions import UnexpectedAPIError
+from deepset_mcp.api.protocols import AsyncClientProtocol
+
+
+async def list_component_families(client: AsyncClientProtocol) -> str:
+    """Lists all Haystack component families that are available on deepset."""
+    haystack_service = client.haystack_service()
+
+    try:
+        response = await haystack_service.get_component_schemas()
+    except UnexpectedAPIError as e:
+        return f"Failed to retrieve component families: {e}"
+
+    components = response["component_schema"]["definitions"]["Components"]
+
+    families = {}
+    for component_def in components.values():
+        component_type = component_def["properties"]["type"]
+        family = component_type["family"]
+        description = component_type.get("family_description", "No description available.")
+        families[family] = description
+
+    if not families:
+        return "No component families found in the response"
+
+    # Format the families into a readable string
+    parts = ["Available Haystack component families:\n"]
+    for family, description in sorted(families.items()):
+        parts.append(f"\n**{family}**\n{description}\n")
+
+    return "\n".join(parts)
